create actios for user info
